requirements.txt
customtkinter
pygame
requests
streamlit
aiogram
python-dotenv